+++ src/Optimisarr.Data/OptimisarrDbContext.cs
@@ -34,6 +34,7 @@ public sealed class OptimisarrDbContext(DbContextOptions<OptimisarrDbContext> op
             entity.Property(library => library.TargetContainer).HasMaxLength(32);
             entity.Property(library => library.ExcludePaths).HasMaxLength(2048);
             entity.Property(library => library.EncoderPreset).HasMaxLength(32);
+            entity.Property(library => library.TargetFolder).HasMaxLength(1024);
         });
 
         modelBuilder.Entity<MediaFile>(entity =>